Use JCUnit8X instead of JCUnit8.
