+++ src/test/java/com/github/dakusui/jcunit8/examples/beforesandafters/BeforeAfter.java
@@ -2,7 +2,7 @@ package com.github.dakusui.jcunit8.examples.beforesandafters;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
 import com.github.dakusui.jcunit8.runners.helpers.ParameterUtils;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.AfterTestCase;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.BeforeTestCase;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
@@ -10,7 +10,7 @@ import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.*;
 import org.junit.runner.RunWith;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class BeforeAfter {
   @ParameterSource
   public Parameter.Factory a() {
