+++ src/test/java/com/github/dakusui/jcunit8/examples/beforesandafters/UnusedParameter.java
@@ -2,13 +2,13 @@ package com.github.dakusui.jcunit8.examples.beforesandafters;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
 import com.github.dakusui.jcunit8.runners.helpers.ParameterUtils;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class UnusedParameter {
   @ParameterSource
   public Parameter.Factory used() {
