+++ src/test/java/com/github/dakusui/jcunit8/examples/config/ConfigExample.java
@@ -3,13 +3,13 @@ package com.github.dakusui.jcunit8.examples.config;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
 import com.github.dakusui.jcunit8.runners.helpers.ParameterUtils;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class ConfigExample {
   @ParameterSource
   public Parameter.Factory<String> platform() {
