+++ src/test/java/com/github/dakusui/jcunit8/examples/parameterizedconstraint/ParameterizedConstraintExample.java
@@ -2,7 +2,7 @@ package com.github.dakusui.jcunit8.examples.parameterizedconstraint;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
 import com.github.dakusui.jcunit8.runners.helpers.ParameterUtils;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.Condition;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.Given;
@@ -12,7 +12,7 @@ import org.junit.runner.RunWith;
 
 import java.util.Arrays;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class ParameterizedConstraintExample {
   @ParameterSource
   public Parameter.Factory a() {
