+++ src/test/java/com/github/dakusui/jcunit8/tests/components/fsm/TurnstileExamples.java
@@ -4,14 +4,14 @@ import com.github.dakusui.jcunit8.factorspace.Parameter;
 import com.github.dakusui.jcunit8.factorspace.fsm.Edge;
 import com.github.dakusui.jcunit8.factorspace.fsm.Player;
 import com.github.dakusui.jcunit8.factorspace.fsm.Scenario;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 public class TurnstileExamples {
-  @RunWith(JCUnit8.class)
+  @RunWith(JCUnit8X.class)
   public static class Normal {
     @ParameterSource
     public Parameter.Factory turnstile() {
@@ -32,7 +32,7 @@ public class TurnstileExamples {
     }
   }
 
-  @RunWith(JCUnit8.class)
+  @RunWith(JCUnit8X.class)
   public static class Broken {
     @ParameterSource
     public Parameter.Factory turnstile() {
