+++ src/test/java/com/github/dakusui/jcunit8/tests/features/seed/SeedFeatureTestBase.java
@@ -2,7 +2,7 @@ package com.github.dakusui.jcunit8.tests.features.seed;
 
 import com.github.dakusui.jcunit.core.tuples.Tuple;
 import com.github.dakusui.jcunit8.factorspace.Parameter;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.Condition;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ConfigureWith;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
@@ -16,7 +16,7 @@ import java.util.List;
 
 import static java.util.Arrays.asList;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class SeedFeatureTestBase {
   static final List<Tuple> testCases = Collections.synchronizedList(new LinkedList<>());
