+++ src/test/java/com/github/dakusui/jcunit8/tests/validation/testresources/UndefinedParameterReferenced.java
@@ -1,14 +1,14 @@
 package com.github.dakusui.jcunit8.tests.validation.testresources;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static java.util.Arrays.asList;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class UndefinedParameterReferenced extends InvalidTestClass {
   /*
    * This method is hidden and errors should be reported.
