+++ src/test/java/com/github/dakusui/jcunit8/tests/validation/testresources/seedfeature/SeedBase.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.jcunit8.tests.validation.testresources.seedfeature;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
@@ -9,7 +9,7 @@ import org.junit.runner.RunWith;
 
 import static java.util.Arrays.asList;
 
-@RunWith(JCUnit8.class)
+@RunWith(JCUnit8X.class)
 public class SeedBase {
   @ParameterSource
   public Parameter.Simple.Factory<String> parameter1() {
